python/requirements.txt
pip==22.3.1
Flask
Flask_Cors
numpy==1.23.5
pillow==9.1.1
opencv-python==4.6.0.66
torch==1.11.0
pandas==1.5.2
tqdm==4.64.1
torchmetrics==0.11.0
lightly==1.2.22
lightning-bolts==0.5.0
pytorch-lightning==1.6.4
einops==0.7.0
timm==0.9.12
scikit-learn==1.2.0
albumentations==1.3.1
tensorboard==2.11.0
Jinja2==3.1.2
cx_Freeze